Resolved an issue with Result widget for geojson with Infinity value
When there is an Infinity as a numebr in the GEOJSON value, QT parser will fail. Thus, the INFINITY value is repalced with inf in qoutes. Also, a condition check was added to check for any parse error so that instead of error and failure of the process result, the failure is handled gracefully.